Refactor Group Data type
