+++ src/lib/types/api/index.ts
@@ -1,0 +1,2 @@
+export * from "./group";
+export * from "./interest";
