+++ src/lib/types/api/interest.ts
@@ -1,0 +1,6 @@
+export interface Interest {
+  id: number;
+  key: string;
+  color: string;
+  emoji: string;
+}
