+++ src/lib/types/index.ts
@@ -2,3 +2,4 @@
 export * from './home';
 export * from './user';
 export * from './shared';
+export * from './api';
